add Cumilative Power and inital queue
For Arduino
added Cumilative Power to the values being sent. Made adjustments so
that each reading sends an individual packet. Also implemented a simple
queue method which is to be upgraded to a sliding window at a later date

For python code
It will now read each packet individually and send an acknowledgement
after reading